# Swagger_Documentation.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Employee Management API",
    "description": "APIs for managing employees",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/employees/{id}": {
      "get": {
        "tags": [
          "Employee"
        ],
        "summary": "Retrieve an employee by ID",
        "description": "Fetches detailed information about a specific employee by their unique ID, including personal and job-related details.",
        "operationId": "getEmployeeById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique ID of the employee to retrieve",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee found and details retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDTO"
                }
              }
            }
          },
          "404": {
            "description": "No employee found with the provided ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Employee"
        ],
        "summary": "Update an existing employee's details",
        "description": "Updates the information of an employee based on their ID. The details provided will overwrite the existing data for that employee.",
        "operationId": "updateEmployee",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the employee whose details need to be updated",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployeeDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Employee details successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid employee data provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Employee not found with the provided ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Employee"
        ],
        "summary": "Delete an employee record",
        "description": "Deletes the employee record identified by the given ID from the system.",
        "operationId": "deleteEmployee",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the employee to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee record successfully deleted"
          },
          "404": {
            "description": "Employee not found with the provided ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/employees": {
      "get": {
        "tags": [
          "Employee"
        ],
        "summary": "Get a list of all employees",
        "description": "Retrieves a comprehensive list of all employees in the system, including their personal and job-related information.",
        "operationId": "getEmployees",
        "responses": {
          "200": {
            "description": "Successfully retrieved the list of all employees",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Employee"
        ],
        "summary": "Create a new employee",
        "description": "Creates a new employee record with the provided details, including personal information and job-related data.",
        "operationId": "createEmployee",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Employee successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid employee data provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/{projectId}/employees/{employeeId}": {
      "delete": {
        "tags": [
          "Project"
        ],
        "summary": "Remove an employee from a project",
        "description": "Removes a specific employee from a project using both the project ID and the employee ID.",
        "operationId": "removeEmployeeFromProject",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of the project from which to remove the employee",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "employeeId",
            "in": "path",
            "description": "ID of the employee to be removed from the project",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee successfully removed from the project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDTO"
                }
              }
            }
          },
          "404": {
            "description": "Project or employee not found for the given IDs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred during input validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "EmployeeDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "departmentName": {
            "type": "string"
          },
          "joinDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "UpdateEmployeeDTO": {
        "type": "object",
        "description": "Updated details of the employee",
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 10,
            "minLength": 10,
            "pattern": "^[0-9]*$"
          },
          "role": {
            "type": "string"
          }
        }
      },
...
      "ProjectDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "deadline": {
            "type": "string",
            "format": "date"
          },
          "departmentName": {
            "type": "string"
          }
        }
      },
...
      "CreateEmployeeDTO": {
        "type": "object",
        "description": "The employee data required for creating a new employee",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "maxLength": 10,
            "minLength": 10,
            "pattern": "^[0-9]*$"
          },
          "role": {
            "type": "string"
          },
          "departmentId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "departmentId",
          "email",
          "name",
          "phone",
          "role"
        ]
      },
...
    }
  }
}